ES-013.1 | Booking&Validation
add validations
