+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/services/order/OrderProceedAfterPaymentService.java
@@ -1,4 +1,4 @@
-package com.wojciechbarwinski.demo.epic_board_games_shop.services;
+package com.wojciechbarwinski.demo.epic_board_games_shop.services.order;
 
 import com.wojciechbarwinski.demo.epic_board_games_shop.legendaryWarehouse.LegendaryWarehousePort;
 import com.wojciechbarwinski.demo.epic_board_games_shop.dtos.OrderDataFromWarehouseDTO;
@@ -7,6 +7,7 @@ import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
 import com.wojciechbarwinski.demo.epic_board_games_shop.entities.OrderStatus;
 import com.wojciechbarwinski.demo.epic_board_games_shop.mappers.MapperFacade;
 import com.wojciechbarwinski.demo.epic_board_games_shop.repositories.OrderRepository;
+import com.wojciechbarwinski.demo.epic_board_games_shop.validations.OrderStatusChangeValidation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -20,11 +21,13 @@ class OrderProceedAfterPaymentService {
     private final MapperFacade mapper;
     private final OrderHelper orderHelper;
     private final LegendaryWarehousePort port;
+    private final OrderStatusChangeValidation orderStatusChangeValidation;
 
     void proceedOrderAfterPayment(String codeId) {
         Long id = Long.parseLong(codeId);// method to read codeId from link as order ID
         Order order = orderHelper.getOrderById(id);
 
+        orderStatusChangeValidation.assertOrderStatusTransitionIsAllowed(order.getOrderStatus(), OrderStatus.PAID);
         order.setOrderStatus(OrderStatus.PAID); //tmp useless
         log.info("Order with id {} was paid", order.getId());
 
@@ -33,6 +36,7 @@ class OrderProceedAfterPaymentService {
 
         try {
             OrderDataFromWarehouseDTO orderDataFromWarehouseDTO = port.sendOrderToWarehouse(orderDataToWarehouseDTO);
+            orderStatusChangeValidation.assertOrderStatusTransitionIsAllowed(order.getOrderStatus(), orderDataFromWarehouseDTO.status());
             order.setOrderStatus(orderDataFromWarehouseDTO.status());
         } catch (Exception e) {
             log.warn("Order with id {} was NOT sending to warehouse successfully", order.getId(), e);
