+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/services/order/OrderReadService.java
@@ -1,31 +1,36 @@
-package com.wojciechbarwinski.demo.epic_board_games_shop.services;
+package com.wojciechbarwinski.demo.epic_board_games_shop.services.order;
 
 import com.wojciechbarwinski.demo.epic_board_games_shop.dtos.OrderResponseDTO;
+import com.wojciechbarwinski.demo.epic_board_games_shop.dtos.OrderSearchRequestDTO;
 import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
-import com.wojciechbarwinski.demo.epic_board_games_shop.entities.OrderStatus;
 import com.wojciechbarwinski.demo.epic_board_games_shop.mappers.MapperFacade;
 import com.wojciechbarwinski.demo.epic_board_games_shop.repositories.OrderRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 
-@Slf4j
 @RequiredArgsConstructor
 @Service
-class OrderProceedAfterConfirmService {
+class OrderReadService {
 
     private final OrderRepository orderRepository;
     private final MapperFacade mapper;
     private final OrderHelper orderHelper;
 
-    OrderResponseDTO proceedOrderAfterConfirm(Long id) {
+    OrderResponseDTO getOrderById(Long id) {
         Order order = orderHelper.getOrderById(id);
 
-        order.setOrderStatus(OrderStatus.CONFIRMED);
-        log.info("Order with id {} was confirm", order.getId());
-        //here will be method that create payment-link and sent it to client mail to pay OR to decline order
-        orderRepository.save(order);
         return mapper.mapOrderToOrderResponseDTO(order);
     }
+
+    List<OrderResponseDTO> getAllOrdersBySearchingData(OrderSearchRequestDTO orderSearchRequestDTO) {
+
+        List<Order> ordersBySearchRequest = orderRepository.findOrdersBySearchRequest(orderSearchRequestDTO);
+
+        return ordersBySearchRequest.stream()
+                .map(mapper::mapOrderToOrderResponseDTO)
+                .toList();
+    }
+
 }
