+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/services/product/ProductIncreaseService.java
@@ -1,4 +1,4 @@
-package com.wojciechbarwinski.demo.epic_board_games_shop.services;
+package com.wojciechbarwinski.demo.epic_board_games_shop.services.product;
 
 import com.wojciechbarwinski.demo.epic_board_games_shop.entities.OrderLine;
 import com.wojciechbarwinski.demo.epic_board_games_shop.repositories.ProductRepository;
