+++ src/test/java/com/wojciechbarwinski/demo/epic_board_games_shop/services/order/OrderRejectServiceTest.java
@@ -1,24 +1,22 @@
-package com.wojciechbarwinski.demo.epic_board_games_shop.services;
+package com.wojciechbarwinski.demo.epic_board_games_shop.services.order;
 
-import com.wojciechbarwinski.demo.epic_board_games_shop.dtos.OrderDataFromWarehouseDTO;
 import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
 import com.wojciechbarwinski.demo.epic_board_games_shop.entities.OrderStatus;
 import com.wojciechbarwinski.demo.epic_board_games_shop.repositories.OrderRepository;
+import com.wojciechbarwinski.demo.epic_board_games_shop.services.product.ProductServicesFacade;
+import com.wojciechbarwinski.demo.epic_board_games_shop.validations.OrderStatusChangeValidation;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class OrderUpdateStatusFromWarehouseServiceTest {
-
+class OrderRejectServiceTest {
 
     @Mock
     private OrderRepository orderRepository;
@@ -26,15 +24,20 @@ class OrderUpdateStatusFromWarehouseServiceTest {
     @Mock
     private OrderHelper orderHelper;
 
+    @Mock
+    private ProductServicesFacade productServicesFacade;
+
+    @Mock
+    private OrderStatusChangeValidation orderStatusChangeValidation;
+
     @InjectMocks
-    private OrderUpdateStatusFromWarehouseService orderUpdateStatusFromWarehouseService;
+    private OrderRejectService orderRejectService;
 
     @Test
-    void shouldUpdateOrderStatusWhenOrderExists() {
+    void shouldSetOrderStatusOnCancelledAfterOrderWasRejected(){
         // given
         Long orderId = 1L;
-        OrderStatus newStatus = OrderStatus.DELIVERED;
-        OrderDataFromWarehouseDTO orderDataDTO = new OrderDataFromWarehouseDTO(orderId, newStatus);
+        OrderStatus statusAfterCancelled = OrderStatus.CANCELLED;
 
         Order order = new Order();
         order.setId(orderId);
@@ -42,14 +45,12 @@ class OrderUpdateStatusFromWarehouseServiceTest {
 
         when(orderHelper.getOrderById(orderId)).thenReturn(order);
 
-        // when
-        orderUpdateStatusFromWarehouseService.updateStatusFromWarehouse(orderDataDTO);
+        //when
+        orderRejectService.rejectOrder(orderId);
 
         // then
         verify(orderHelper).getOrderById(orderId);
         verify(orderRepository).save(order);
-        assertEquals(newStatus, order.getOrderStatus());
+        assertEquals(statusAfterCancelled, order.getOrderStatus());
     }
-
-
 }
